wip updates to e2e tests
