+++ package.json
@@ -14,8 +14,8 @@
   ],
   "main": "bin/main.js",
   "scripts": {
-    "test-js": "node --experimental-fetch --experimental-vm-modules node_modules/.bin/jest --verbose",
-    "lint-js": "eslint up.ts test/ --ext .js,.jsx,.ts,.tsx",
+    "test": "node --experimental-fetch --experimental-vm-modules node_modules/.bin/jest --verbose",
+    "lint": "eslint up.ts test/ --ext .js,.jsx,.ts,.tsx",
     "prepublishOnly": "npm run build",
     "prettier": "prettier '**/*.{ts,json,md}' --check",
     "prettier:fix": "npm run prettier -- --write",
