+++ test/e2e.test.js
@@ -1,3 +1,5 @@
+import { spawnSync } from "node:child_process";
+import { join } from "node:path";
 import { providers } from "ethers";
 import { jest } from "@jest/globals";
 import path from "path";
@@ -13,13 +15,11 @@ const __dirname = path.resolve(path.dirname(""));
 // TODO: we were using these tests to check the validator's OAS spec via
 // copy copying the file during local tableland startup. Now that is a dev
 // product, these kind of tests need to be separated
-// e.g.
-/*  spawnSync("mkdir", ["./tmp"]);
-    spawnSync("cp", [
-      join(this.validatorDir, "tableland-openapi-spec.yaml"),
-      "./tmp",
-    ]);
-*/
+spawnSync("mkdir", ["./tmp"]);
+spawnSync("cp", [
+  join(__dirname, "../go-tableland", "tableland-openapi-spec.yaml"),
+  "./tmp",
+]);
 
 
 // These tests take a bit longer than normal since we are usually waiting for blocks to finalize etc...
@@ -28,6 +28,16 @@ const accounts = getAccounts();
 
 // NOTE: these tests require the a local Tableland is already running
 describe("Validator, Chain, and SDK work end to end", function () {
+  test("just do a single read", async function () {
+    const signer = accounts[1];
+    const tableland = await getTableland(signer);
+
+    const data = await tableland.read(
+      `SELECT * FROM test_create_read_31337_68;`
+    );
+    expect(data).toEqual([]);
+  });
+
   test("Create a table that can be read from", async function () {
     const signer = accounts[1];
 
@@ -44,7 +54,7 @@ describe("Validator, Chain, and SDK work end to end", function () {
     const data = await tableland.read(
       `SELECT * FROM ${prefix}_${chainId}_${tableId};`
     );
-    await expect(data.rows).toEqual([]);
+    expect(data).toEqual([]);
   });
 
   test("Create a table that can be written to", async function () {
@@ -67,7 +77,7 @@ describe("Validator, Chain, and SDK work end to end", function () {
     const data = await tableland.read(`SELECT * FROM ${queryableName};`);
 
     await expect(typeof writeRes.hash).toEqual("string");
-    await expect(data.rows).toEqual([["tree", "aspen"]]);
+    await expect(data).toEqual([{keyy: "tree", val: "aspen"}]);
   });
 
   test("Table cannot be written to unless caller is allowed", async function () {
@@ -82,7 +92,7 @@ describe("Validator, Chain, and SDK work end to end", function () {
     const queryableName = `${prefix}_${chainId}_${tableId}`;
 
     const data = await tableland.read(`SELECT * FROM ${queryableName};`);
-    await expect(data.rows).toEqual([]);
+    await expect(data).toEqual([]);
 
     const signer2 = accounts[2];
     const tableland2 = await getTableland(signer2);
@@ -94,7 +104,7 @@ describe("Validator, Chain, and SDK work end to end", function () {
     );
 
     const data2 = await tableland2.read(`SELECT * FROM ${queryableName};`);
-    await expect(data2.rows).toEqual([]);
+    await expect(data2).toEqual([]);
   });
 
   test("Create a table can have a row deleted", async function () {
@@ -359,75 +369,3 @@ describe("Validator, Chain, and SDK work end to end", function () {
   });
 
 });
-
-describe("Validator gateway server", function () {
-  let token, transactionHash, tableHash, schemaTableId;
-  beforeAll(async function () {
-    // TODO: split openapi spec tests and js tests into different files and npm commands,
-    //       then `npm test` can run everything.
-    const signer0 = accounts[0];
-    const tableland0 = await getTableland(signer0);
-    await tableland0.siwe();
-
-    // We can"t use the Validator's Wallet to create tables because the Validator's nonce tracking will get out of sync
-    const signer1 = accounts[1];
-    const tableland1 = await getTableland(signer1);
-
-    const prefix = "test_transaction";
-    const { txnHash, tableId } = await tableland1.create("keyy TEXT, val TEXT", { prefix });
-
-    const { tableId: tableId2 } = await tableland1.create("a INT PRIMARY KEY, CHECK (a > 0)", { prefix: "test_schema_route" });
-    schemaTableId = tableId2;
-
-    const { structureHash } = await tableland1.hash("a INT PRIMARY KEY", { prefix: "test_schema_route" });
-    tableHash = structureHash;
-
-    const chainId = 31337;
-
-    const data = await tableland1.read(`SELECT * FROM ${prefix}_${chainId}_${tableId};`);
-    await expect(data.rows).toEqual([]);
-
-    // We need the token and a transaction hash for a transaction on the Hardhat chain,
-    // to run the tests for the openapi spec file so we hoist them here..
-    token = tableland0.token.token;
-    transactionHash = txnHash;
-  });
-
-  const tests = loadSpecTestData(path.join(__dirname, "tmp", "tableland-openapi-spec.yaml"));
-
-  test.each(tests)("$name", async function (_test) {
-    const payload = {
-      method: _test.methodName.toUpperCase(),
-      headers: {
-        "Content-Type": "application/json",
-        Authorization: `Bearer ${token}`,
-      }
-    };
-
-    const routeTemplateData = {
-      chainID: 31337,
-      id: 1,
-      address: "0xf39fd6e51aad88f6f4ce6ab8827279cfffb92266", // Hardhat #1
-      readStatement: "SELECT * FROM healthbot_31337_1",
-      tableName: `test_schema_route_31337_${schemaTableId}`,
-      hash: tableHash
-    };
-
-    // Cannot have a body on a GET/HEAD request
-    if (_test.body) {
-      // For some of the example requests we need to inject values for the chain tests are using
-      if (_test.body.params && _test.body.params[0].txn_hash) _test.body.params[0].txn_hash = transactionHash;
-      payload.body = JSON.stringify(_test.body);
-    }
-
-    const route = _test.route(routeTemplateData)
-    const res = await fetch(`${_test.host}${route}`, payload);
-
-    expect(typeof _test.response).not.toEqual("undefined");
-
-    if (route === "/rpc") return await testRpcResponse(res, _test);
-    await testHttpResponse(res, _test.response)
-  });
-});
-
-
